requirements-dev.txt
coverage==4.0.3
flake8==2.5.4
mock==1.3.0
nose==1.3.7
pep8==1.6.2
rednose>=0.4.3
vcrpy>=1.7.4
